pyproject.toml
[project]
name = "tavilyagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.2.5",
    "langchain-tavily>=0.2.0",
    "langchain[anthropic]>=0.3.25",
    "langgraph>=0.4.8",
    "loguru>=0.7.3",
    "uvicorn>=0.34.3",
]
